+++ My new project/categoriesform.cs
@@ -55,13 +55,20 @@ namespace My_new_project
 
         private void categoriesform_Load(object sender, EventArgs e)
         {
-            populate(); 
+            populate();
+            CatID.Hide();
+            label3.Hide();
+            label9.Hide();
+            label10.Hide();
+           
         }
 
 
         private void guna2DataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
             label8.Hide();
+            label9.Show();
+            label10.Show();
             CatID.Text = CatDGV.SelectedRows[0].Cells[0].Value.ToString();
             Catname.Text = CatDGV.SelectedRows[0].Cells[1].Value.ToString();
             CatDes.Text = CatDGV.SelectedRows[0].Cells[2].Value.ToString();
@@ -77,16 +84,26 @@ namespace My_new_project
                     MessageBox.Show("Select the category to delete");
                 }
                 else {
+                    DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete the employee information ?", "Some Title", MessageBoxButtons.YesNo);
+                    if (dialogResult == DialogResult.Yes)
+                    {
 
-                    Con.Open();
-                    String query = "delete from CategoryTable where CatID = "+CatID.Text+"";
-                    SqlCommand cmd = new SqlCommand(query, Con);
-                    cmd.ExecuteNonQuery();
-                    MessageBox.Show("Category is deleted Successfully");
-                    Con.Close();
-                    populate();
-                    cleartext();
-                    populate();
+                        Con.Open();
+                        String query = "delete from CategoryTable where CatID = " + CatID.Text + "";
+                        SqlCommand cmd = new SqlCommand(query, Con);
+                        cmd.ExecuteNonQuery();
+                        MessageBox.Show("Category is deleted Successfully");
+                        Con.Close();
+                        populate();
+                        cleartext();
+
+                        label9.Hide();
+                        label10.Hide();
+                    }
+                    else {
+                        label9.Hide();
+                        label10.Hide();
+                    }
                 }
             }
             catch (Exception ex)
@@ -103,6 +120,9 @@ namespace My_new_project
         private void label11_Click(object sender, EventArgs e)
         {
             cleartext();
+            label9.Hide();
+            label10.Hide();
+            label8.Show();
         }
 
         private void label9_Click(object sender, EventArgs e)
@@ -124,6 +144,8 @@ namespace My_new_project
                     populate();
                     cleartext();
                     populate();
+                    label9.Hide();
+                    label10.Hide();
                 }
             }
             catch (Exception ex)
@@ -172,6 +194,11 @@ namespace My_new_project
         private void CatID_MaskInputRejected(object sender, MaskInputRejectedEventArgs e)
         {
 
+        }
+
+        private void label6_Click(object sender, EventArgs e)
+        {
+            Application.Exit();
         }
     }
 }
